Use en_US.UTF-8 locale instead of C.UTF-8 for fedora
C.UTF-8 is not supported by glibc, but some distributions like fedora
and debian patch it in. Using it causes problems, for instance ssh
sessions will carry the locale configuration, resulting in an invalid
locale being selected if the target server does not support it (like
RHEL and CentOS). The end result is terminal weirdness.

Using en_US.UTF-8 is a more sensible default. It's also what a regular
fedora 30 installation defaults to (at least the livecd does).